+++ examples/color.rs
@@ -12,11 +12,9 @@ use embedded_graphics_simulator::{
     OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent, Window,
 };
 use embedded_menu::{
-    interaction::simulator::Simulator,
-    items::{select::SelectValue, NavigationItem, Select},
-    selection_indicator::style::rectangle::Rectangle as RectangleIndicator,
-    theme::Theme,
-    Menu, MenuStyle,
+    interaction::simulator::Simulator, items::menu_item::SelectValue,
+    selection_indicator::style::rectangle::Rectangle as RectangleIndicator, theme::Theme, Menu,
+    MenuStyle,
 };
 
 #[derive(Copy, Clone, PartialEq)]
@@ -35,7 +33,7 @@ impl SelectValue for TestEnum {
         }
     }
 
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             TestEnum::A => "A",
             TestEnum::B => "AB",
@@ -73,17 +71,17 @@ fn main() -> Result<(), core::convert::Infallible> {
                 esc_value: (),
             }),
     )
-    .add_item(NavigationItem::new("Foo", ()).with_marker(">"))
-    .add_item(Select::new("Check this", false))
-    .add_item(Select::new("Check this", false))
-    .add_item(Select::new("Check this too", false))
+    .add_item("Foo", ">", |_| ())
+    .add_item("Check this", false, |_| ())
+    .add_item("Check this", TestEnum::A, |_| ())
+    .add_item("Check this too", false, |_| ())
     .build();
 
     let output_settings = OutputSettingsBuilder::new().scale(4).build();
     let mut window = Window::new("Menu demonstration w/color", &output_settings);
 
     'running: loop {
-        let mut display: SimulatorDisplay<Rgb888> = SimulatorDisplay::new(Size::new(128, 64));
+        let mut display = SimulatorDisplay::new(Size::new(128, 64));
         let mut sub = display.cropped(&Rectangle::new(Point::new(16, 16), Size::new(96, 34)));
         menu.update(&sub);
         menu.draw(&mut sub).unwrap();