+++ examples/color_single_touch.rs
@@ -12,11 +12,8 @@ use embedded_graphics_simulator::{
     sdl2::Keycode, OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent, Window,
 };
 use embedded_menu::{
-    interaction::single_touch::SingleTouch,
-    items::{select::SelectValue, NavigationItem, Select},
-    selection_indicator::style::AnimatedTriangle,
-    theme::Theme,
-    Menu, MenuStyle,
+    interaction::single_touch::SingleTouch, items::menu_item::SelectValue,
+    selection_indicator::style::AnimatedTriangle, theme::Theme, Menu, MenuStyle,
 };
 
 #[derive(Copy, Clone, PartialEq)]
@@ -35,7 +32,7 @@ impl SelectValue for TestEnum {
         }
     }
 
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             TestEnum::A => "A",
             TestEnum::B => "AB",
@@ -74,10 +71,10 @@ fn main() -> Result<(), core::convert::Infallible> {
                 max_time: 100,
             }),
     )
-    .add_item(NavigationItem::new("Foo", ()).with_marker(">"))
-    .add_item(Select::new("Check this", false))
-    .add_item(Select::new("Check this", false))
-    .add_item(Select::new("Check this too", false))
+    .add_item("Foo", ">", |_| ())
+    .add_item("Check this", false, |_| ())
+    .add_item("Check this", false, |_| ())
+    .add_item("Check this too", false, |_| ())
     .build();
 
     let output_settings = OutputSettingsBuilder::new().scale(4).build();
@@ -85,7 +82,7 @@ fn main() -> Result<(), core::convert::Infallible> {
 
     let mut space_pressed = false;
     'running: loop {
-        let mut display: SimulatorDisplay<Rgb888> = SimulatorDisplay::new(Size::new(128, 64));
+        let mut display = SimulatorDisplay::new(Size::new(128, 64));
         let mut sub = display.cropped(&Rectangle::new(Point::new(16, 16), Size::new(96, 34)));
         menu.update(&sub);
         menu.draw(&mut sub).unwrap();