+++ examples/font.rs
@@ -4,7 +4,6 @@
 
 use embedded_graphics::{
     mono_font::{ascii::FONT_8X13_BOLD, iso_8859_1::FONT_6X10},
-    pixelcolor::BinaryColor,
     prelude::Size,
     Drawable,
 };
@@ -12,9 +11,7 @@ use embedded_graphics_simulator::{
     BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent, Window,
 };
 use embedded_menu::{
-    interaction::simulator::Simulator,
-    items::{select::SelectValue, NavigationItem, Select},
-    Menu, MenuStyle,
+    interaction::simulator::Simulator, items::menu_item::SelectValue, Menu, MenuStyle,
 };
 
 #[derive(Copy, Clone, PartialEq)]
@@ -33,7 +30,7 @@ impl SelectValue for TestEnum {
         }
     }
 
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             TestEnum::A => "A",
             TestEnum::B => "AB",
@@ -53,12 +50,10 @@ fn main() -> Result<(), core::convert::Infallible> {
                 esc_value: (),
             }),
     )
-    .add_item(
-        NavigationItem::new("Nav item", ()).with_marker("»"), // not part of the ASCII font
-    )
-    .add_item(Select::new("Checkbox", true))
-    .add_item(Select::new("Other checkbox", false))
-    .add_item(Select::new("Multiple options long", TestEnum::A))
+    .add_item("Nav item", "»", |_| ()) // » is not part of the ASCII font
+    .add_item("Checkbox", true, |_| ())
+    .add_item("Other checkbox", false, |_| ())
+    .add_item("Multiple options long", TestEnum::A, |_| ())
     .build();
 
     let output_settings = OutputSettingsBuilder::new()
@@ -67,7 +62,7 @@ fn main() -> Result<(), core::convert::Infallible> {
     let mut window = Window::new("Menu demonstration", &output_settings);
 
     'running: loop {
-        let mut display: SimulatorDisplay<BinaryColor> = SimulatorDisplay::new(Size::new(128, 64));
+        let mut display = SimulatorDisplay::new(Size::new(128, 64));
         menu.update(&display);
         menu.draw(&mut display).unwrap();
         window.update(&display);