+++ examples/owned.rs
@@ -2,13 +2,13 @@
 //!
 //! Navigate using up/down arrows, interact using the Enter key
 
-use embedded_graphics::{pixelcolor::BinaryColor, prelude::Size, Drawable};
+use embedded_graphics::{prelude::Size, Drawable};
 use embedded_graphics_simulator::{
     BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent, Window,
 };
 use embedded_menu::{
     interaction::simulator::Simulator,
-    items::{select::SelectValue, NavigationItem, Select},
+    items::{menu_item::SelectValue, MenuItem},
     Menu, MenuStyle,
 };
 
@@ -28,7 +28,7 @@ impl SelectValue for TestEnum {
         }
     }
 
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             TestEnum::A => "A",
             TestEnum::B => "AB",
@@ -39,9 +39,9 @@ impl SelectValue for TestEnum {
 
 fn main() -> Result<(), core::convert::Infallible> {
     // Use a generator to create owned strings that we give to the menu.
-    let items: Vec<Select<_, _, _>> = (1..10)
+    let items: Vec<MenuItem<_, _, _, true>> = (1..10)
         .map(|i| format!("Item {}", i))
-        .map(|i| Select::new(i, false))
+        .map(|i| MenuItem::new(i, false))
         .collect();
 
     let mut menu = Menu::with_style(
@@ -51,8 +51,8 @@ fn main() -> Result<(), core::convert::Infallible> {
             esc_value: (),
         }),
     )
-    .add_item(NavigationItem::new("Foo", ()).with_marker(">"))
-    .add_items(items)
+    .add_item("Foo", ">", |_| ())
+    .add_menu_items(items)
     .build();
 
     let output_settings = OutputSettingsBuilder::new()
@@ -61,7 +61,7 @@ fn main() -> Result<(), core::convert::Infallible> {
     let mut window = Window::new("Menu demonstration", &output_settings);
 
     'running: loop {
-        let mut display: SimulatorDisplay<BinaryColor> = SimulatorDisplay::new(Size::new(128, 64));
+        let mut display = SimulatorDisplay::new(Size::new(128, 64));
         menu.update(&display);
         menu.draw(&mut display).unwrap();
         window.update(&display);