+++ examples/scrolling.rs
@@ -5,9 +5,7 @@ use embedded_graphics_simulator::{
     BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent, Window,
 };
 use embedded_menu::{
-    interaction::simulator::Simulator,
-    items::{select::SelectValue, NavigationItem, Select},
-    Menu, MenuStyle,
+    interaction::simulator::Simulator, items::menu_item::SelectValue, Menu, MenuStyle,
 };
 
 #[derive(Copy, Clone, PartialEq)]
@@ -26,7 +24,7 @@ impl SelectValue for TestEnum {
         }
     }
 
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             TestEnum::A => "A",
             TestEnum::B => "AB",
@@ -44,19 +42,19 @@ fn main() -> Result<(), core::convert::Infallible> {
         .with_animated_selection_indicator(10);
 
     let mut menu = Menu::with_style("Menu", style)
-        .add_item(NavigationItem::new("Foo", ()).with_marker(">"))
-        .add_item(Select::new("Check this", false))
-        .add_item(Select::new("Check this", false))
-        .add_item(Select::new("Check this too", TestEnum::A))
-        .add_item(Select::new("Check this too", TestEnum::A))
-        .add_item(Select::new("Check this", true))
-        .add_item(Select::new("Check this too", true))
-        .add_item(Select::new("Check this too", TestEnum::A))
-        .add_item(Select::new("Check this", false))
-        .add_item(Select::new("Check this too", true))
-        .add_item(NavigationItem::new("Foo", ()).with_marker(">"))
-        .add_item(Select::new("Check this", false))
-        .add_item(Select::new("Check this too", TestEnum::A))
+        .add_item("Foo", ">", |_| ())
+        .add_item("Check this", false, |_| ())
+        .add_item("Check this", false, |_| ())
+        .add_item("Check this too", TestEnum::A, |_| ())
+        .add_item("Check this too", TestEnum::A, |_| ())
+        .add_item("Check this", true, |_| ())
+        .add_item("Check this too", true, |_| ())
+        .add_item("Check this too", TestEnum::A, |_| ())
+        .add_item("Check this", false, |_| ())
+        .add_item("Check this too", true, |_| ())
+        .add_item("Foo", "<-", |_| ())
+        .add_item("Check this", false, |_| ())
+        .add_item("Check this too", TestEnum::A, |_| ())
         .build();
 
     let output_settings = OutputSettingsBuilder::new()
@@ -65,7 +63,7 @@ fn main() -> Result<(), core::convert::Infallible> {
     let mut window = Window::new("Menu demonstration", &output_settings);
 
     'running: loop {
-        let mut display: SimulatorDisplay<BinaryColor> = SimulatorDisplay::new(Size::new(128, 64));
+        let mut display = SimulatorDisplay::new(Size::new(128, 64));
         menu.update(&display);
         menu.draw(&mut display).unwrap();
         window.update(&display);