+++ examples/scrolling_single_touch.rs
@@ -3,16 +3,14 @@
 //! Navigate using only the spacebar. Short(ish) press moves on to the next item, long press activates.
 //! Watch the animated selection indicator fill up. Long press is registered as the bar reaches full width.
 
-use embedded_graphics::{pixelcolor::BinaryColor, prelude::Size, Drawable};
+use embedded_graphics::{prelude::Size, Drawable};
 use embedded_graphics_simulator::{
     sdl2::Keycode, BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent,
     Window,
 };
 use embedded_menu::{
-    interaction::single_touch::SingleTouch,
-    items::{select::SelectValue, NavigationItem, Select},
-    selection_indicator::style::animated_triangle::AnimatedTriangle,
-    Menu, MenuStyle,
+    interaction::single_touch::SingleTouch, items::menu_item::SelectValue,
+    selection_indicator::style::animated_triangle::AnimatedTriangle, Menu, MenuStyle,
 };
 
 #[derive(Copy, Clone, PartialEq)]
@@ -31,7 +29,7 @@ impl SelectValue for TestEnum {
         }
     }
 
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             TestEnum::A => "A",
             TestEnum::B => "AB",
@@ -51,19 +49,19 @@ fn main() -> Result<(), core::convert::Infallible> {
         .with_animated_selection_indicator(10);
 
     let mut menu = Menu::with_style("Menu with even longer title", style)
-        .add_item(NavigationItem::new("Foo", ()).with_marker(">"))
-        .add_item(Select::new("Check this", false))
-        .add_item(Select::new("Check this", false))
-        .add_item(Select::new("Check this too", TestEnum::A))
-        .add_item(Select::new("Check this too", TestEnum::A))
-        .add_item(Select::new("Check this", true))
-        .add_item(Select::new("Check this too", true))
-        .add_item(Select::new("Check this too", TestEnum::A))
-        .add_item(Select::new("Check this", false))
-        .add_item(Select::new("Check this too", true))
-        .add_item(NavigationItem::new("Foo", ()).with_marker(">"))
-        .add_item(Select::new("Check this", false))
-        .add_item(Select::new("Check this too", TestEnum::A))
+        .add_item("Foo", ">", |_| ())
+        .add_item("Check this", false, |_| ())
+        .add_item("Check this", false, |_| ())
+        .add_item("Check this too", TestEnum::A, |_| ())
+        .add_item("Check this too", TestEnum::A, |_| ())
+        .add_item("Check this", true, |_| ())
+        .add_item("Check this too", true, |_| ())
+        .add_item("Check this too", TestEnum::A, |_| ())
+        .add_item("Check this", false, |_| ())
+        .add_item("Check this too", true, |_| ())
+        .add_item("Foo", "<-", |_| ())
+        .add_item("Check this", false, |_| ())
+        .add_item("Check this too", TestEnum::A, |_| ())
         .build();
 
     let output_settings = OutputSettingsBuilder::new()
@@ -73,7 +71,7 @@ fn main() -> Result<(), core::convert::Infallible> {
 
     let mut space_pressed = false;
     'running: loop {
-        let mut display: SimulatorDisplay<BinaryColor> = SimulatorDisplay::new(Size::new(128, 64));
+        let mut display = SimulatorDisplay::new(Size::new(128, 64));
         menu.update(&display);
         menu.draw(&mut display).unwrap();
         window.update(&display);