+++ examples/scrolling_slice.rs
@@ -1,12 +1,12 @@
 //! Run using `cargo run --example scrolling --target x86_64-pc-windows-msvc` --features=simulator
 
-use embedded_graphics::{pixelcolor::BinaryColor, prelude::Size, Drawable};
+use embedded_graphics::{prelude::Size, Drawable};
 use embedded_graphics_simulator::{
     BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent, Window,
 };
 use embedded_menu::{
     interaction::simulator::Simulator,
-    items::{select::SelectValue, NavigationItem, Select},
+    items::{menu_item::SelectValue, MenuItem},
     Menu, MenuStyle,
 };
 
@@ -26,7 +26,7 @@ impl SelectValue for TestEnum {
         }
     }
 
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             TestEnum::A => "A",
             TestEnum::B => "AB",
@@ -44,21 +44,21 @@ fn main() -> Result<(), core::convert::Infallible> {
         .with_animated_selection_indicator(10);
 
     let mut selects1 = [
-        Select::new("Check this 1", false),
-        Select::new("Check this 2", false),
+        MenuItem::new("Check this 1", false),
+        MenuItem::new("Check this 2", false),
     ];
     let mut selects2 = [
-        Select::new("Check this 3", true),
-        Select::new("Check this 4", true),
+        MenuItem::new("Check this 3", true),
+        MenuItem::new("Check this 4", true),
     ];
 
     let mut menu = Menu::with_style("Menu", style)
-        .add_item(NavigationItem::new("Foo", ()).with_marker(">"))
-        .add_items(&mut selects1)
-        .add_items(&mut selects2)
-        .add_item(NavigationItem::new("Foo", ()).with_marker(">"))
-        .add_item(Select::new("Check this", false))
-        .add_item(Select::new("Check this too", TestEnum::A))
+        .add_item("Foo", ">", |_| ())
+        .add_menu_items(&mut selects1)
+        .add_menu_items(&mut selects2)
+        .add_item("Foo", "<-", |_| ())
+        .add_item("Check this", false, |_| ())
+        .add_item("Check this too", TestEnum::A, |_| ())
         .build();
 
     let output_settings = OutputSettingsBuilder::new()
@@ -67,7 +67,7 @@ fn main() -> Result<(), core::convert::Infallible> {
     let mut window = Window::new("Menu demonstration", &output_settings);
 
     'running: loop {
-        let mut display: SimulatorDisplay<BinaryColor> = SimulatorDisplay::new(Size::new(128, 64));
+        let mut display = SimulatorDisplay::new(Size::new(128, 64));
         menu.update(&display);
         menu.draw(&mut display).unwrap();
         window.update(&display);