+++ examples/small.rs
@@ -3,7 +3,6 @@
 //! Navigate using up/down arrows, interact using the Enter key
 
 use embedded_graphics::{
-    pixelcolor::BinaryColor,
     prelude::{DrawTargetExt, Point, Size},
     primitives::Rectangle,
     Drawable,
@@ -12,9 +11,7 @@ use embedded_graphics_simulator::{
     BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent, Window,
 };
 use embedded_menu::{
-    interaction::simulator::Simulator,
-    items::{select::SelectValue, NavigationItem, Select},
-    Menu, MenuStyle,
+    interaction::simulator::Simulator, items::menu_item::SelectValue, Menu, MenuStyle,
 };
 
 #[derive(Copy, Clone, PartialEq)]
@@ -33,7 +30,7 @@ impl SelectValue for TestEnum {
         }
     }
 
-    fn name(&self) -> &'static str {
+    fn marker(&self) -> &'static str {
         match self {
             TestEnum::A => "A",
             TestEnum::B => "AB",
@@ -50,10 +47,10 @@ fn main() -> Result<(), core::convert::Infallible> {
             esc_value: (),
         }),
     )
-    .add_item(NavigationItem::new("Foo", ()).with_marker(">"))
-    .add_item(Select::new("Check this", false))
-    .add_item(Select::new("Check this", false))
-    .add_item(Select::new("Check this too", false))
+    .add_item("Foo", ">", |_| ())
+    .add_item("Check this", false, |_| ())
+    .add_item("Check this", false, |_| ())
+    .add_item("Check this too", false, |_| ())
     .build();
 
     let output_settings = OutputSettingsBuilder::new()
@@ -62,7 +59,7 @@ fn main() -> Result<(), core::convert::Infallible> {
     let mut window = Window::new("Menu demonstration", &output_settings);
 
     'running: loop {
-        let mut display: SimulatorDisplay<BinaryColor> = SimulatorDisplay::new(Size::new(128, 64));
+        let mut display = SimulatorDisplay::new(Size::new(128, 64));
         let mut sub = display.cropped(&Rectangle::new(Point::new(16, 16), Size::new(96, 34)));
         menu.update(&sub);
         menu.draw(&mut sub).unwrap();